fix(style): hidden new thread btn on new thread page
