+++ src/renderer/src/views/ChatTabView.vue
@@ -13,25 +13,6 @@
           :session-id="pageRouter.chatSessionId"
         />
       </template>
-
-      <Transition name="collapsed-new-chat-button">
-        <div
-          v-if="showCollapsedNewChatButton"
-          class="pointer-events-none absolute inset-x-0 top-0 z-30 h-12"
-        >
-          <Button
-            variant="ghost"
-            size="icon"
-            data-testid="collapsed-new-chat-button"
-            class="collapsed-new-chat-button pointer-events-auto absolute left-4 top-2.5 h-7 w-7 text-muted-foreground hover:text-foreground"
-            :title="t('common.newChat')"
-            :aria-label="t('common.newChat')"
-            @click="handleCollapsedNewChat"
-          >
-            <Icon icon="lucide:plus" class="h-4 w-4" />
-          </Button>
-        </div>
-      </Transition>
     </div>
 
     <ChatSidePanel
@@ -42,10 +23,7 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted, ref } from 'vue'
-import { Icon } from '@iconify/vue'
-import { Button } from '@shadcn/components/ui/button'
-import { useI18n } from 'vue-i18n'
+import { onMounted, ref } from 'vue'
 import ChatSidePanel from '@/components/sidepanel/ChatSidePanel.vue'
 import NewThreadPage from '@/pages/NewThreadPage.vue'
 import ChatPage from '@/pages/ChatPage.vue'
@@ -53,24 +31,13 @@ import AgentWelcomePage from '@/pages/AgentWelcomePage.vue'
 import { usePageRouterStore } from '@/stores/ui/pageRouter'
 import { useSessionStore } from '@/stores/ui/session'
 import { useAgentStore } from '@/stores/ui/agent'
-import { useSidebarStore } from '@/stores/ui/sidebar'
 import { useProjectStore } from '@/stores/ui/project'
 
-const { t } = useI18n()
 const pageRouter = usePageRouterStore()
 const sessionStore = useSessionStore()
 const agentStore = useAgentStore()
-const sidebarStore = useSidebarStore()
 const projectStore = useProjectStore()
 const isReady = ref(false)
-const showCollapsedNewChatButton = computed(
-  () =>
-    isReady.value && sidebarStore.collapsed && Boolean(sessionStore.newConversationTargetAgentId)
-)
-
-const handleCollapsedNewChat = () => {
-  void sessionStore.startNewConversation({ refresh: true })
-}
 
 onMounted(async () => {
   try {
@@ -104,29 +71,5 @@ onMounted(async () => {
 
 ::-webkit-scrollbar-thumb:hover {
   background: #9ca3af80;
-}
-
-.collapsed-new-chat-button-enter-active,
-.collapsed-new-chat-button-leave-active {
-  transition:
-    opacity 200ms ease-out,
-    transform 200ms ease-out;
-}
-
-.collapsed-new-chat-button-enter-from,
-.collapsed-new-chat-button-leave-to {
-  opacity: 0;
-  transform: translateX(-10px);
-}
-
-.collapsed-new-chat-button-enter-to,
-.collapsed-new-chat-button-leave-from {
-  opacity: 1;
-  transform: translateX(0);
-}
-
-.collapsed-new-chat-button {
-  -webkit-app-region: no-drag;
-  pointer-events: auto;
 }
 </style>
